Use ReadOnlyMemory<char> for framework cache and add unit test for ReadOnlyMemoryCharComparerOrdinal
